backend/requirements.txt
# backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.4
psycopg2-binary>=2.9,<2.10
redis>=5.0,<5.1
celery>=5.4,<5.5
gunicorn>=22.0,<23.0
python-decouple>=3.8,<3.9
pytest>=8.0,<9.0
pytest-django>=4.8,<4.9
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<3.4
python-telegram-bot>=21.0,<22.0
pytest-asyncio>=0.23,<1.0
google-generativeai>=0.8.0,<1.0
groq>=0.11.0,<1.0
